feat(j): migrate checkout.session.completed to registry (J-01c-2)
The 509-line checkout.session.completed case in route.ts was the last
large block in the legacy switch. Keeping it there meant six business
flows (course purchase, advisor top-up, advisor featured, listing
activation, consultation booking, sponsored placement) lived in a
single untestable 1200-line file — one bad edit could silently break
all six. Moving it to a dedicated handler file means each flow is
visible, independently testable, and can be exercised against a mock
WebhookContext without spinning up the full route.

Verified callers / context:
  - app/api/stripe/webhook/route.ts — POST handler, service-role admin
    client via createAdminClient(); stripe via getStripe(). These are
    now passed through WebhookContext so the handler never reaches for
    them directly (keeps the test seam predictable per J-01 design).
  - lib/stripe-webhook/handlers/index.ts — registers the handler via
    registerHandler("checkout.session.completed", ...) before
    dispatchEvent runs.
  - lib/stripe-webhook/lib/email.ts — email helpers consumed by the
    handler (buildCourseReceiptEmail, buildConsultationConfirmationEmail,
    emailWrapper, sendTransactionalEmail).

Behaviour preserved byte-for-byte:
  - Idempotent advisor_credit_topup: skips and returns done when
    topup row status is already "completed".
  - Course purchase rollback: deletes purchase row + fires admin alert
    when creator revenue insert fails.
  - Sponsored placement 23505 race-tolerance: loser treats
    unique-violation as success; metadata-incomplete path fires admin
    alert and continues.
  - Lazy invoice-URL fetch: ctx.stripe.invoices.retrieve(session.invoice)
    is non-fatal, result flows into email + booking row.

Idempotency claim: the outer idempotency wrapper (stripe_webhook_events
table, processing → done state machine) in route.ts still applies.
Within the handler, advisor_credit_topup has its own topup_id check;
sponsored_placement has a select-existing + 23505 guard.

Rollback: revert this commit; the checkout.session.completed case body
is preserved in git history at the commit before this one on branch
claude/audit-remediation/j-stripe-webhook.

Queue item: J-01c-2
Audit section: §J "Stripe webhook completeness"
Refs: #221
